requirements.txt
cassandra-driver
pyasyncore
requests
matplotlib
